requirements.txt
psutil==6.0.0
py-cpuinfo==9.0.0
GPUtil==1.4.0
setuptools>=65.5.1
